src/main/java/com/bouygtel/jenkins/redmine/RedmineGroupDetails.java
//
// Copyright (C) 2013 Bouygues Telecom (cynoffic@bouyguestelecom.fr)
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//         http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
//

package com.bouygtel.jenkins.redmine;

import hudson.security.GroupDetails;

import org.acegisecurity.GrantedAuthority;

/**
 * Redmine implementation for {@link GroupDetails}
 * 
 * @author Cyrille Nofficial
 */
public class RedmineGroupDetails extends GroupDetails implements GrantedAuthority {

    private static final long serialVersionUID = -5253590554711997820L;
    private final String name;

    /**
     * Constructeur
     * 
     * @param name
     */
    public RedmineGroupDetails(String name) {
        this.name = name;
    }

    /**
     * {@inheritDoc}
     * 
     * @see hudson.security.GroupDetails#getName()
     */
    @Override
    public String getName() {
        return name;
    }

    /**
     * {@inheritDoc}
     * 
     * @see org.acegisecurity.GrantedAuthority#getAuthority()
     */
    @Override
    public String getAuthority() {
        return name;
    }

    /**
     * {@inheritDoc}
     * 
     * @see java.lang.Object#hashCode()
     */
    @Override
    public int hashCode() {
        final int prime = 31;
        int result = 1;
        result = prime * result + ((name == null) ? 0 : name.hashCode());
        return result;
    }

    /**
     * {@inheritDoc}
     * 
     * @see java.lang.Object#equals(java.lang.Object)
     */
    @Override
    public boolean equals(Object obj) {
        if (this == obj) return true;
        if (obj == null) return false;
        if (getClass() != obj.getClass()) return false;
        RedmineGroupDetails other = (RedmineGroupDetails) obj;
        if (name == null) {
            if (other.name != null) return false;
        } else if (!name.equals(other.name)) return false;
        return true;
    }

    /**
     * {@inheritDoc}
     * 
     * @see java.lang.Object#toString()
     */
    @Override
    public String toString() {
        StringBuilder builder = new StringBuilder();
        builder.append("RedmineGroupDetails [name=");
        builder.append(name);
        builder.append("]");
        return builder.toString();
    }

}
